Group shader variables with the same name
So they only appear once in the UI. Usually they're just separate
instances of the same variable in HLSL (e.g. multiple compilations with
the same shader with different defines). Even if they're conceptually
different, there's no way to distinguish which is which (they have the
same name!) so it's not a good idea. If you want different variables,
give them different names.