Disable assembly loading hook in IronPython
We don't need it, and it causes lots of spurious exception thrown
messages in the debug out when trying to load assemblies using our
'project:' path prefix (assembly loading doesn't go through the PAL).

(Note that the exceptions are a problem as such, they just clutter up
the log, hiding important messages).