Re-compile shaders when the include lookup changes
Previously, if the lookup function itself changed, but the contents of
all files stayed the same, you'd get a cache hit, so the effect of the
function change was ignored.

This manifests itself in the custom mip generation, because changing the
name of the file to use from script changed the result of the (custom)
include lookup function, but nothing else, so the result of changing the
script was not seen.