WIP deferred render context
Defer actual rendering rather than running during the script execution.
The idea is so that we can collect required resources/shaders and
create/compile them lazily. This lets us only compile used shaders
without having to interrupt rendering to do so (which would be bad UX,
as it wouldn't be async). It does none of this yet, this is just a
skeleton!