Deferred shader compilation
Still synchronous and serial, but framework is in place. List of shaders
is built up during script execution and compiled once execution is
complete. Bound shader vars are similarly just stored and actual binding
is done afterwards. This is not an error checking problem as we already
silently ignored missing shader vars.