Allow user var access in main script body
Script is automatically re-executed when changing a variable accessed
thusly. Instead of re-running the script from source, it re-executes the
same compiled output from the original run. This reduces the chance that
the user has changed the script and the variable disappears, which would
be confusing, or alternatively that the user is half way through editing
the script, and it no longer compiles.